dashboard/requirements.txt
dash 
dash-renderer 
dash-html-components 
dash-core-components 
jinja2
psycopg2-binary
